Tests: share provider registration helpers
